+++ src/js/types.js
@@ -50,26 +50,24 @@ var types = {
         validate: function( x ) {
             var pattern = /([0-9]{4})([\-]|[\/])([0-9]{2})([\-]|[\/])([0-9]{2})/;
             var segments = pattern.exec( x );
-
-            return ( segments && segments.length === 6 ) ? ( new Date( Number( segments[ 1 ] ) + '-' + Number( segments[ 3 ] ) + '-' + Number( segments[ 5 ] ) ).toString() !== 'Invalid Date' ) : false;
+            if ( segments && segments.length === 6 ) {
+                var year = Number( segments[ 1 ] );
+                var month = Number( segments[ 3 ] ) - 1;
+                var day = Number( segments[ 5 ] );
+                var date = new Date( year, month, day );
+                // Do not approve automatic JavaScript conversion of invalid dates such as 2017-12-32
+                return date.getFullYear() === year && date.getMonth() === month && date.getDate() === day;
+            }
+            return false;
         },
         convert: function( x ) {
-            var pattern = /([0-9]{4})([\-]|[\/])([0-9]{2})([\-]|[\/])([0-9]{2})/;
-            var segments;
             var date;
 
             if ( utils.isNumber( x ) ) {
                 // The XPath expression "2012-01-01" + 2 returns a number of days in XPath.
                 date = new Date( x * 24 * 60 * 60 * 1000 );
             } else {
-                date = new Date( x );
-                if ( date.toString() === 'Invalid Date' ) {
-                    segments = pattern.exec( x );
-                    //this code is really only meant for the Rhino and PhantomJS engines, in browsers it may never be reached
-                    if ( segments && Number( segments[ 1 ] ) > 0 && Number( segments[ 3 ] ) >= 0 && Number( segments[ 3 ] ) < 12 && Number( segments[ 5 ] ) < 32 ) {
-                        date = new Date( Number( segments[ 1 ] ), ( Number( segments[ 3 ] ) - 1 ), Number( segments[ 5 ] ) );
-                    }
-                }
+                date = this.validate( x ) ? new Date( x ) : 'Invalid Date';
             }
 
             return date.toString() === 'Invalid Date' ?
@@ -78,51 +76,71 @@ var types = {
     },
     'datetime': {
         validate: function( x ) {
-            //the second part builds in some tolerance for slightly-off dates provides as defaults (e.g.: 2013-05-31T07:00-02)
-            return ( new Date( x.toString() ).toString() !== 'Invalid Date' || new Date( this.convert( x.toString() ) ).toString() !== 'Invalid Date' );
+            var parts = x.split( 'T' );
+            if ( parts.length === 2 ) {
+                return types.date.validate( parts[ 0 ] ) && types.time.validate( parts[ 1 ], false );
+            }
+
+            return false;
         },
         convert: function( x ) {
-            var date;
-            var patternCorrect = /([0-9]{4}\-[0-9]{2}\-[0-9]{2})([T]|[\s])([0-9]){2}:([0-9]){2}([0-9:.]*)(\+|\-)([0-9]{2}):([0-9]{2})$/;
-            var patternAlmostCorrect = /([0-9]{4}\-[0-9]{2}\-[0-9]{2})([T]|[\s])([0-9]){2}:([0-9]){2}([0-9:.]*)(\+|\-)([0-9]{2})$/;
-
+            var date = 'Invalid Date';
+            var parts = x.split( 'T' );
             if ( utils.isNumber( x ) ) {
                 // The XPath expression "2012-01-01T01:02:03+01:00" + 2 returns a number of days in XPath.
                 date = new Date( x * 24 * 60 * 60 * 1000 );
-            } else if ( new Date( x ).toString() !== 'Invalid Date' && patternCorrect.test( x ) ) {
-                // Do not risk changing the time zone by calling toISOLocalString()
-                return x;
-            } else if ( new Date( x ).toString() === 'Invalid Date' && patternAlmostCorrect.test( x ) ) {
-                // Do not risk changing the time zone by calling toISOLocalString()
-                return x + ':00';
+            } else if ( /[0-9]T[0-9]/.test( x ) && parts.length === 2 ) {
+                var convertedDate = types.date.convert( parts[ 0 ] );
+                // The milliseconds are optional for datetime (and shouldn't be added)
+                var convertedTime = types.time.convert( parts[ 1 ], false );
+                if ( convertedDate && convertedTime ) {
+                    return convertedDate + 'T' + convertedTime;
+                }
             } else {
+                // XPath returns a string like this: "Mon, 08 Jan 2018 18:46:36 GMT".
+                // Until we fix this in the XPath evaluator to return an ISOLocal string, we should deal with this.
                 date = new Date( x );
             }
 
-            return ( date.toString() !== 'Invalid Date' ) ? date.toISOLocalString() : '';
+            return date.toString() !== 'Invalid Date' ? date.toISOLocalString() : '';
         }
     },
     'time': {
-        validate: function( x ) {
+        // Note that it's okay if the validate function is stricter than the spec,
+        // (for timezone offset), as long as the convertor automatically converts
+        // to a valid time.
+        validate: function( x, requireMillis ) {
             var m = x.match( /^(\d\d):(\d\d):(\d\d)\.\d\d\d(\+|-)(\d\d):(\d\d)$/ );
+
+            requireMillis = typeof requireMillis !== 'boolean' ? true : requireMillis;
+
+            if ( !m && !requireMillis ) {
+                m = x.match( /^(\d\d):(\d\d):(\d\d)(\+|-)(\d\d):(\d\d)$/ );
+            }
 
             if ( !m ) {
                 return false;
             }
 
-            // no need to convert to numbers since we now they are number strings
+            // no need to convert to numbers since we know they are number strings
             return m[ 1 ] < 24 && m[ 1 ] >= 0 &&
                 m[ 2 ] < 60 && m[ 2 ] >= 0 &&
                 m[ 3 ] < 60 && m[ 3 ] >= 0 &&
                 m[ 5 ] < 24 && m[ 5 ] >= 0 && // this could be tighter
                 m[ 6 ] < 60 && m[ 6 ] >= 0; // this is probably either 0 or 30
         },
-        convert: function( x ) {
+        convert: function( x, requireMillis ) {
             var date;
             var o = {};
             var parts;
+            var time;
+            var secs;
+            var tz;
             var offset;
-            var timeAppearsCorrect = /^[0-9]{2}:[0-9]{2}(:[0-9.]*)?$/;
+            var noMillis = false;
+            var timeAppearsCorrect = /^[0-9]{1,2}:[0-9]{1,2}(:[0-9.]*)?/;
+
+            requireMillis = typeof requireMillis !== 'boolean' ? true : requireMillis;
 
             if ( !timeAppearsCorrect.test( x ) ) {
                 // An XPath expression would return a datetime string since there is no way to request a timeValue.
@@ -130,32 +148,41 @@ var types = {
                 date = new Date( x );
                 if ( date.toString() !== 'Invalid Date' ) {
                     x = date.getHours() + ':' + date.getMinutes() + ':' + date.getSeconds() + '.' + date.getMilliseconds() + date.getTimezoneOffsetAsTime();
+                } else {
+                    return '';
                 }
             }
 
-            parts = x.toString().split( ':' );
+            if ( /^([0-9]{2}:[0-9]{2}:[0-9]{2})([+-][0-9]{2}.*)$/.test( x ) ) {
+                noMillis = true;
+            }
 
-            if ( parts.length < 2 ) {
+            parts = x.toString().split( /(\+|-|Z)/ );
+            // We're using a 'capturing group' here, so the + or - is included!.
+            if ( parts.length < 1 ) {
                 return '';
             }
 
-            o.hours = parts[ 0 ].pad( 2 );
-            o.minutes = parts[ 1 ].pad( 2 );
-            parts = ( parts[ 2 ] || '00' ).split( '.' );
-            o.seconds = parts[ 0 ].pad( 2 );
-            parts = ( parts[ 1 ] || '000' ).split( /(\+|-)/ );
-            o.milliseconds = parts[ 0 ].pad( 3 );
-            if ( parts.length === 1 ) {
+            time = parts[ 0 ].split( ':' );
+            tz = parts[ 2 ] ? [ parts[ 1 ] ].concat( parts[ 2 ].split( ':' ) ) : ( parts[ 1 ] === 'Z' ? [ '+', '00', '00' ] : [] );
+
+            o.hours = time[ 0 ].pad( 2 );
+            o.minutes = time[ 1 ].pad( 2 );
+
+            secs = time[ 2 ] ? time[ 2 ].split( '.' ) : [ '00' ];
+
+            o.seconds = secs[ 0 ];
+            o.milliseconds = secs[ 1 ] || ( requireMillis ? '000' : undefined );
+
+            if ( tz.length === 0 ) {
                 offset = new Date().getTimezoneOffsetAsTime();
-                parts[ 1 ] = offset.substring( 0, 1 );
-                parts[ 2 ] = offset.substring( 1 );
+            } else {
+                offset = tz[ 0 ] + tz[ 1 ].pad( 2 ) + ':' + ( tz[ 2 ] ? tz[ 2 ].pad( 2 ) : '00' );
             }
-            o.offsetDirection = parts[ 1 ];
-            o.offsetTime = parts[ 2 ].length > 2 ? parts[ 2 ] : parts[ 2 ].pad( 2 ) + ':00';
 
-            x = o.hours + ':' + o.minutes + ':' + o.seconds + '.' + o.milliseconds + o.offsetDirection + o.offsetTime;
+            x = o.hours + ':' + o.minutes + ':' + o.seconds + ( o.milliseconds ? '.' + o.milliseconds : '' ) + offset;
 
-            return this.validate( x ) ? x : '';
+            return this.validate( x, requireMillis ) ? x : '';
         }
     },
     'barcode': {